move: widget 파일 정리
