+++ lib/pages/login/login_page.dart
@@ -1,4 +1,4 @@
-import 'package:dino_shop/pages/login/widgets/english_only.dart';
+import 'package:dino_shop/pages/login/widgets/log_in_box.dart';
 import 'package:dino_shop/theme.dart';
 import 'package:flutter/material.dart';
 
@@ -28,36 +28,10 @@ class LoginPage extends StatefulWidget {
 }
 
 class _LoginPageState extends State<LoginPage> {
-  final _emailController = TextEditingController();
-  final _passwordController = TextEditingController();
-
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      body: Padding(
-        padding: EdgeInsets.all(30),
-        child: Column(
-          children: [
-            Image.asset('assets/images/application/logo.png'),
-            TextFormField(
-                controller: _emailController,
-                decoration: InputDecoration(labelText: 'EMAIL'),
-                keyboardType: TextInputType.emailAddress,
-                inputFormatters: [EnglishOnly()]),
-            TextFormField(
-              controller: _passwordController,
-              obscureText: true,
-              decoration: InputDecoration(labelText: 'PASSWORD'),
-            ),
-            ElevatedButton(
-                onPressed: () {
-                  print('이메일: ${_emailController.text}');
-                  print('비밀번호: ${_passwordController}');
-                },
-                child: Text('LOGIN'))
-          ],
-        ),
-      ),
+      body: LogInBox(),
     );
   }
 }
